requirements.txt
Flask==3.0.0
SQLAlchemy>=2.0.36
psycopg[binary]
kafka-python-ng==2.2.2
PyJWT==2.8.0
pydantic>=2.9
python-dotenv==1.0.0
werkzeug==3.0.1
requests
